+++ core/administration.go
@@ -1546,7 +1546,7 @@ func (app *Application) getUserByExternalID(externalID string) (*model.User, err
 	return user, nil
 }
 
-func (app *Application) createAction(providerID string, userID string, encryptedKey string, encryptedBlob string) (*model.CTest, error) {
+func (app *Application) createAction(current model.User, group string, providerID string, userID string, encryptedKey string, encryptedBlob string) (*model.CTest, error) {
 	//1. create a ctest
 	item, user, err := app.storage.CreateAdminCTest(providerID, userID, encryptedKey, encryptedBlob, false, nil)
 	if err != nil {
@@ -1576,6 +1576,12 @@ func (app *Application) createAction(providerID string, userID string, encrypted
 		data["click_action"] = "FLUTTER_NOTIFICATION_CLICK"
 		app.messaging.SendNotificationMessage(userData.FCMTokens, "COVID-19", "You have received a COVID-19 update", data)
 	}(user.UUID)
+
+	//audit
+	userIdentifier, userInfo := current.GetLogData()
+	lData := []AuditDataEntry{{Key: "providerID", Value: providerID}, {Key: "userID", Value: userID},
+		{Key: "encryptedKey", Value: encryptedKey}, {Key: "encryptedBlob", Value: encryptedBlob}}
+	defer app.audit.LogCreateEvent(userIdentifier, userInfo, group, "action", item.ID, lData)
 
 	return item, nil
 }